Added Class functionality to sensors and estimator.

+++ Microcontroller_Code/autoPilot/include/SDs.h
@@ -2,7 +2,7 @@
 #define SD_H
 
 bool setupSD();
-void writeData(double *state, double *servoInput, double *receiverInput, bool *autoMode, int *auxMode);
+void writeData(float *state, float *servoInput, float *receiverInput, bool *autoMode, int *auxMode);
 void openSD();
 void closeSD();
