+++ Microcontroller_Code/autoPilot/include/servos.h
@@ -3,10 +3,10 @@
 
 void setup_Servos();
 
-void writeServosAngle(double *inputAngle);
+void writeServosAngle(float *inputAngle);
 
 void writeServosPWM(volatile unsigned long *inputPWM);
 
-void getRCSignalAngle(double *input, bool *autoMode, int *auxMode);
+void getRCSignalAngle(float *input, bool *autoMode, int *auxMode);
 
 #endif
