+++ Microcontroller_Code/autoPilot/include/stabilize.h
@@ -1,7 +1,6 @@
 #ifndef STABILIZE_H
 #define STABILIZE_H
 
-void stabilize(double *receiverInput, double *servoInput, double *state);
-
+void stabilize(float *receiverInput, float *servoInput, float *state);
 
 #endif
