+++ Microcontroller_Code/autoPilot/include/test.h
@@ -1,4 +1,4 @@
 #ifndef PROJECT__TEST__H
 #define PROJECT__TEST__H
-double volume(double breadth, double height, double length);
+float volume(float breadth, float height, float length);
 #endif
