+++ Microcontroller_Code/autoPilot/src/stabilize.cpp
@@ -2,14 +2,13 @@
 #include <servos.h>
 
 //  --- Stabilize --- //
-void stabilize(double *receiverInput, double *servoInput, double *state)
+void stabilize(float *receiverInput, float *servoInput, float *state)
 {
     // Controls
     servoInput[0] = 0;  // Throttle
-    servoInput[1] = 90;  // Elevator
-    servoInput[2] = 90;  // Aileron
-    servoInput[3] = 90;  // Rudder
-
+    servoInput[1] = 90; // Elevator
+    servoInput[2] = 90; // Aileron
+    servoInput[3] = 90; // Rudder
 
     // Actuate Measurements
     writeServosAngle(&servoInput[0]);
