+++ Microcontroller_Code/autoPilot/src/test.cpp
@@ -1,7 +1,8 @@
 #include "test.h"
 #include <math.h>
-double volume(double breadth, double height, double length) {
-   double volume_result;
-   volume_result = breadth*height*length;
+float volume(float breadth, float height, float length)
+{
+   float volume_result;
+   volume_result = breadth * height * length;
    return volume_result;
 }
